+++ lib/postgres_pgmq.dart
@@ -11,14 +11,6 @@
 /// await pgmq.createQueue('emails');
 /// final id = await pgmq.send('emails', {'to': 'a@example.com'});
 /// final msg = await pgmq.pop('emails');
-/// ```dart
-/// import 'package:postgres/postgres.dart';
-/// import 'package:postgres_pgmq/postgres_pgmq.dart';
-///
-/// final pgmq = Pgmq(connection);
-/// await pgmq.createQueue('emails');
-/// final id = await pgmq.send('emails', {'to': 'a@example.com'});
-/// final msg = await pgmq.pop('emails');
 /// ```
 library;